Add private filesystem module
- Separate filesystem.rs into filesystem/* modules
